adding documentation
all functions now have a docstring.

the TODO Orgfile has been updated to reflect the planned roadmap.

the README has been updated to clarify terminology used in the
documentation and to reflect the change in orientation from
a "secure BBS" to a "simple BBS".

the couch views documentation has been updated with documentation
explaining why each view exists and what purpose it serves.